pyproject.toml
[tool.poetry]
name = "ri3521-backend"
version = "0.1.0"
description = ""
authors = ["GoodWeather0322 <a82521112@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
pymysql = "^1.1.1"
python-multipart = "^0.0.10"
aiofiles = "^24.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.29.1"
cz-conventional-gitmoji = "^0.3.3"
sqlite-web = "^0.6.4"
pytest = "^8.3.3"
black = "^24.8.0"


[tool.commitizen]
name = "cz_gitmoji"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
